Add 'all' section to configuration file
The task load_yaml_config now allows for a cascading configuration style similar to symfony yaml configuration files.
A file loaded with this task must have at least an 'all' section which contains the default configuration (which can be empty). Any other section on the same level as 'all' is considered an environment that can be specified in the env parameter.
First the parameters in 'all' are loaded and are then updated by the parameters in the specified section.
If both 'all' and for example 'dev' had the same key, the value specified in 'dev' would override the one in 'all'. This allows for a base configuration that might contain parameters that are common to all environments with a more specialized configuration based on the deployment target.